Move global menu to the toolbar
relevant comment:
https://github.com/cockpit-project/cockpit-files/pull/567#issuecomment-2337581719
